WISDjango/requirements.txt
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
billiard==4.1.0
black==23.3.0
celery==5.3.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
Django==4.2
djangorestframework==3.14.0
install==1.3.5
kombu==5.3.1
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
platformdirs==3.5.1
prompt-toolkit==3.0.38
psycopg2-binary==2.9.6
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
redis==4.5.5
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
uWSGI==2.0.21
vine==5.0.0
wcwidth==0.2.6
